# perf: Get exact diff nodes during incremental sync