--- a/.changeset/tiny-chicken-marry.md
+++ b/.changeset/tiny-chicken-marry.md
@@ -1,0 +1,5 @@
+---
+"@farcaster/hubble": patch
+---
+
+perf: Improve incremental sync performance