## Motivation

Improve performance of incremental sync

## Change Summary

- During incremental sync, we were accidentally fetching messages that we already had
- Make sure we get the exact node/message that is missing instead of getting everything from a leaf. 
- Remove some duplicate logging

## Merge Checklist

_Choose all relevant options below by adding an `x` now or at any time before submitting for review_

- [X] PR title adheres to the [conventional commits](https://www.conventionalcommits.org/en/v1.0.0/) standard
- [X] PR has a [changeset](https://github.com/farcasterxyz/hub-monorepo/blob/main/CONTRIBUTING.md#35-adding-changesets)
- [X] PR has been tagged with a change label(s) (i.e. documentation, feature, bugfix, or chore)
- [X] PR does not require changes to the [protocol](https://github.com/farcasterxyz/protocol)
- [X] All [commits have been signed](https://github.com/farcasterxyz/hub-monorepo/blob/main/CONTRIBUTING.md#22-signing-commits)

Stats:
Time drops from 17s -> 9s, and data fetched drops from 11kb / call -> 0.3kb / call
```
Fast Divergence Measurement
With optimization (9s 0.3kb / call)
Total Time

          | Duration (s)
------------------------
Wall Time |            9

Latencies (ms)

                 Method | Count | Min | Max |  Avg  | Median
------------------------------------------------------------
getSyncMetadataByPrefix |   371 |  41 |  75 | 44.50 |     42
  getAllSyncIdsByPrefix |    79 |  41 |  59 | 46.04 |     45
getAllMessagesBySyncIds |    22 |  43 |  65 | 56.32 |     63
          mergeMessages |    22 |   7 | 132 | 79.50 |    102

Data Fetched (bytes)

                 Method | Count | Objects |  Total | Min |  Max |    Avg | Median
---------------------------------------------------------------------------------
getSyncMetadataByPrefix |   371 |     371 | 175.6K |  78 |  800 | 473.40 |    787
  getAllSyncIdsByPrefix |    79 |     111 |   8.8K |  80 |  369 | 111.56 |     82
getAllMessagesBySyncIds |    22 |      54 |  20.8K | 346 | 1.9K | 947.50 |    348
          mergeMessages |    22 |      54 |      0 |   0 |    0 |      0 |      0

Without optimisation (17s, 11kb / call)

Total Time

          | Duration (s)
------------------------
Wall Time |           17

Latencies (ms)

                 Method | Count | Min | Max |   Avg  | Median
-------------------------------------------------------------
getSyncMetadataByPrefix |   344 |  43 | 375 |  88.85 |     53
  getAllSyncIdsByPrefix |    61 |  44 | 100 |  62.20 |     44
getAllMessagesBySyncIds |    44 |  98 | 101 |  99.50 |     99
          mergeMessages |    44 | 215 | 282 | 246.75 |    257

Data Fetched (bytes)

                 Method | Count | Objects |  Total |  Min |   Max |    Avg | Median
-----------------------------------------------------------------------------------
getSyncMetadataByPrefix |   344 |     344 | 168.6K |   78 |   800 | 490.20 |    764
  getAllSyncIdsByPrefix |    61 |     156 |  11.8K |   80 |  2.3K | 193.61 |     81
getAllMessagesBySyncIds |    44 |      54 |  37.1K | 6.7K | 11.6K |   9.3K |  11.6K
          mergeMessages |    44 |      54 |      0 |    0 |     0 |      0 |      0
```

<!-- start pr-codex -->

---

## PR-Codex overview
### Detailed summary
- Improved incremental sync performance in the `MerkleTrie` class by skipping unload if there are no pending DB updates.
- Added a test case to ensure that only the exact missing message is fetched during syncing.
- Updated log levels and messages for better clarity and debugging.
- Added a method to retrieve the number of connected peers in the `SyncEngine` class.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->